move addZoweSession to Profiels class
