+++ __tests__/__unit__/extension.unit.test.ts
@@ -1029,6 +1029,7 @@ describe("Extension Unit Tests", () => {
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
                     return {
+                        addZoweSession: jest.fn(),
                         allProfiles: [profileOne, {name: "secondName"}],
                         defaultProfile: profileOne,
                         validProfile: ValidProfileEnum.VALID,
@@ -1062,26 +1063,24 @@ describe("Extension Unit Tests", () => {
             createQuickPickMock(qpItem, entered);
 
 
-            await utils.addZoweSession(testTree);
+            await Profiles.getInstance().addZoweSession(testTree);
             expect(showInformationMessage.mock.calls[0][0]).toEqual("Profile Name was not supplied. Operation Cancelled");
         });
 
         it("Testing that addSession with supplied profile name", async () => {
             const entered = undefined;
-            const addSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
             showInputBox.mockReturnValueOnce("fake");
-            await utils.addZoweSession(testTree);
-            expect(addSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
         it("Testing that addSession with existing profile", async () => {
             const entered = "";
-            const addSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
@@ -1090,25 +1089,23 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testTree);
-            expect(addSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
 
         it("Testing that addSession with supplied resolveQuickPickHelper", async () => {
             const entered = "fake";
-            const addSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testTree);
-            expect(addSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
         it("Testing that addSession with undefined profile", async () => {
             const entered = "";
-            const addSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
@@ -1117,15 +1114,14 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testTree);
-            expect(addSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
 
         it("Testing that addSession if createNewConnection is invalid", async () => {
             const entered = "fake";
-            const addSession = jest.spyOn(utils, "addZoweSession");
 
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
@@ -1143,14 +1139,13 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testTree);
-            expect(addSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
         it("Testing that addSession if listProfile is invalid", async () => {
             const entered = "fake";
-            const addSession = jest.spyOn(utils, "addZoweSession");
 
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
@@ -1168,8 +1163,8 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testTree);
-            expect(addSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
     });
@@ -2651,6 +2646,7 @@ describe("Extension Unit Tests", () => {
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
                     return {
+                        addZoweSession: jest.fn(),
                         allProfiles: [{name: "firstName"}, {name: "secondName"}],
                         defaultProfile: {name: "firstName"},
                         createNewConnection: jest.fn(()=>{
@@ -2682,26 +2678,24 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testTree);
+            await Profiles.getInstance().addZoweSession(testTree);
             expect(showInformationMessage.mock.calls[0][0]).toEqual("Profile Name was not supplied. Operation Cancelled");
         });
 
         it("Testing that addZoweSession with supplied profile name", async () => {
             const entered = undefined;
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
             showInputBox.mockReturnValueOnce("fake");
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
         it("Testing that addZoweSession with theia", async () => {
             const entered = "";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
             Object.defineProperty(globals, "ISTHEIA", { get: () => true });
 
             // Assert edge condition user cancels the input path box
@@ -2711,15 +2705,14 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
             Object.defineProperty(globals, "ISTHEIA", { get: () => false });
         });
 
         it("Testing that addZoweSession with theia fails if no choice", async () => {
             const entered = null;
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
             Object.defineProperty(globals, "ISTHEIA", { get: () => true });
 
             // Assert edge condition user cancels the input path box
@@ -2729,8 +2722,8 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
             expect(showInformationMessage).toHaveBeenCalled();
 
             Object.defineProperty(globals, "ISTHEIA", { get: () => false });
@@ -2738,7 +2731,6 @@ describe("Extension Unit Tests", () => {
 
         it("Testing that addZoweSession with existing profile", async () => {
             const entered = "";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
@@ -2747,25 +2739,23 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
 
         it("Testing that addZoweSession with supplied resolveQuickPickHelper", async () => {
             const entered = "fake";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
         it("Testing that addZoweSession with undefined profile", async () => {
             const entered = "";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
@@ -2774,15 +2764,14 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
 
         it("Testing that addZoweSession if createNewConnection is invalid", async () => {
             const entered = "fake";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
@@ -2800,14 +2789,13 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
 
         it("Testing that addZoweSession if listProfile is invalid", async () => {
             const entered = "fake";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
@@ -2825,8 +2813,8 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testUSSTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testUSSTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
 
         });
     });
@@ -3468,6 +3456,7 @@ describe("Extension Unit Tests", () => {
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
                     return {
+                        addZoweSession: jest.fn(),
                         allProfiles: [{name: "firstName", profile: {user:undefined, password: undefined}}, {name: "secondName"}],
                         defaultProfile: {name: "firstName"},
                         validProfile: ValidProfileEnum.VALID,
@@ -3497,26 +3486,23 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testJobsTree);
+            await Profiles.getInstance().addZoweSession(testJobsTree);
             expect(showInformationMessage.mock.calls[0][0]).toEqual("Profile Name was not supplied. Operation Cancelled");
         });
 
         it("Testing that addJobsSession with supplied profile name", async () => {
             const entered = undefined;
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
             showInputBox.mockReturnValueOnce("fake");
-            await utils.addZoweSession(testJobsTree);
-            expect(addZoweSession).toHaveBeenCalled();
-
+            await Profiles.getInstance().addZoweSession(testJobsTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
 
         it("Testing that addJobsSession with existing profile", async () => {
             const entered = "";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
@@ -3525,25 +3511,22 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testJobsTree);
-            expect(addZoweSession).toHaveBeenCalled();
+            await Profiles.getInstance().addZoweSession(testJobsTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
 
         it("Testing that addJobsSession with supplied resolveQuickPickHelper", async () => {
             const entered = "fake";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testJobsTree);
-            expect(addZoweSession).toHaveBeenCalled();
-
+            await Profiles.getInstance().addZoweSession(testJobsTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
 
         it("Testing that addJobsSession with undefined profile", async () => {
             const entered = "";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
@@ -3552,15 +3535,12 @@ describe("Extension Unit Tests", () => {
                 () => Promise.resolve(createQuickPick())
             );
 
-            await utils.addZoweSession(testJobsTree);
-            expect(addZoweSession).toHaveBeenCalled();
-
+            await Profiles.getInstance().addZoweSession(testJobsTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
-
 
         it("Testing that addJobsSession if createNewConnection is invalid", async () => {
             const entered = "fake";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
@@ -3578,14 +3558,12 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testJobsTree);
-            expect(addZoweSession).toHaveBeenCalled();
-
+            await Profiles.getInstance().addZoweSession(testJobsTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
 
         it("Testing that addJobsSession if listProfile is invalid", async () => {
             const entered = "fake";
-            const addZoweSession = jest.spyOn(utils, "addZoweSession");
 
             Object.defineProperty(Profiles, "getInstance", {
                 value: jest.fn(() => {
@@ -3603,9 +3581,8 @@ describe("Extension Unit Tests", () => {
             // Assert edge condition user cancels the input path box
             createQuickPickMock(qpItem, entered);
 
-            await utils.addZoweSession(testJobsTree);
-            expect(addZoweSession).toHaveBeenCalled();
-
+            await Profiles.getInstance().addZoweSession(testJobsTree);
+            expect(Profiles.getInstance().addZoweSession).toHaveBeenCalled();
         });
     });
