+++ i18n/sample/src/Profiles.i18n.json
@@ -20,5 +20,13 @@
     "promptcredentials.invalidusername": "Please enter your z/OS username. Operation Cancelled",
     "promptcredentials.option.prompt.password.placeholder": "Password",
     "promptcredentials.option.prompt.password": "Enter a password for the connection",
-    "promptcredentials.invalidpassword": "Please enter your z/OS password. Operation Cancelled"
+    "promptcredentials.invalidpassword": "Please enter your z/OS password. Operation Cancelled",
+    "addZoweSession.log.debug.selectProfile": "User selected profile ",
+    "addZoweSession.log.debug.cancelledSelection": "User cancelled profile selection",
+    "addSession.quickPickOption": "Choose \"Create new...\" to define a new profile or select an existing profile to Add to the USS Explorer",
+    "enterPattern.pattern": "No selection made.",
+    "createNewConnection.option.prompt.profileName.placeholder": "Connection Name",
+    "createNewConnection.option.prompt.profileName": "Enter a name for the connection",
+    "createNewConnection.enterprofileName": "Profile Name was not supplied. Operation Cancelled",
+    "addSession.log.debug.createNewProfile": "User created a new profile"
 }
