+++ i18n/sample/src/utils.i18n.json
@@ -2,14 +2,6 @@
     "errorHandling.invalid.credentials": "Invalid Credentials. Please ensure the username and password for ",
     "errorHandling.invalid.credentials2": " are valid or this may lead to a lock-out.",
     "deactivate.error": "Unable to delete temporary folder. ",
-    "addZoweSession.log.debug.selectProfile": "User selected profile ",
-    "addZoweSession.log.debug.cancelledSelection": "User cancelled profile selection",
-    "addSession.quickPickOption": "Choose \"Create new...\" to define a new profile or select an existing profile to Add to the USS Explorer",
-    "enterPattern.pattern": "No selection made.",
-    "createNewConnection.option.prompt.profileName.placeholder": "Connection Name",
-    "createNewConnection.option.prompt.profileName": "Enter a name for the connection",
-    "createNewConnection.enterprofileName": "Profile Name was not supplied. Operation Cancelled",
-    "addSession.log.debug.createNewProfile": "User created a new profile",
     "moveTempFolder.error": "Error encountered when creating temporary folder! ",
     "profile.init.read.imperative": "Unable to read imperative file. ",
     "initialize.module.load": "Credentials not managed, unable to load security file: "
