+++ src/utils.ts
@@ -17,12 +17,6 @@ import * as path from "path";
 import { moveSync } from "fs-extra";
 import { Profiles } from "./Profiles";
 import { ImperativeConfig } from "@zowe/imperative";
-import { DatasetTree } from "./dataset/DatasetTree";
-import { ZosJobsProvider } from "./job/ZosJobsProvider";
-import { USSTree } from "./uss/USSTree";
-import { IZoweTree } from "./api/IZoweTree";
-import { IZoweDatasetTreeNode, IZoweJobTreeNode, IZoweUSSTreeNode } from "./api/IZoweTreeNode";
-import { ZoweExplorerApiRegister } from "./api/ZoweExplorerApiRegister";
 
 import * as nls from "vscode-nls";
 const localize = nls.config({messageFormat: nls.MessageFormat.file})();
@@ -135,160 +129,6 @@ export async function cleanTempDir() {
       cleanDir(globals.ZOWETEMPFOLDER);
   } catch (err) {
       vscode.window.showErrorMessage(localize("deactivate.error", "Unable to delete temporary folder. ") + err);
-  }
-}
-
-/**
- * Adds a new Profile to the provided treeview by clicking the 'Plus' button and
- * selecting which profile you would like to add from the drop-down that appears.
- * The profiles that are in the tree view already will not appear in the drop-down.
- *
- * @export
- * @param {USSTree} zoweNodeProvider - either the USS, MVS, JES tree
- */
-export async function addZoweSession(zoweNodeProvider: IZoweTree<IZoweDatasetTreeNode|IZoweUSSTreeNode|IZoweJobTreeNode>) {
-  const allProfiles = Profiles.getInstance().allProfiles;
-
-  // Get all profiles
-  let profileNamesList = allProfiles.map((profile) => {
-      return profile.name;
-  });
-
-  // Filter to list of the APIs available for current tree explorer
-  profileNamesList = filterProfileNames(zoweNodeProvider, profileNamesList);
-
-  // Select a profile regardless of whether we are in Theia or not
-  // IFF this function is made into a class, sharing information would be simpler between functions. Also functions will be properly protected
-  const { chosenProfile, quickpickValue } = await selectProfile(profileNamesList);
-
-  if (chosenProfile === "") {
-    // Create profile
-    await createProfile(zoweNodeProvider, quickpickValue);
-  } else if (chosenProfile) {
-      globals.LOG.debug(localize("addZoweSession.log.debug.selectProfile", "User selected profile ") + chosenProfile);
-      await zoweNodeProvider.addSession(chosenProfile);
-  } else {
-      globals.LOG.debug(localize("addZoweSession.log.debug.cancelledSelection", "User cancelled profile selection"));
-  }
-}
-
-/**
- * Filter the list of profile names based on APIs available for current tree explorer
- *
- * @param zoweNodeProvider - either the USS, MVS, JES tree
- * @param profileNamesList - List of all profile names
- */
-function filterProfileNames(zoweNodeProvider: IZoweTree<IZoweDatasetTreeNode|IZoweUSSTreeNode|IZoweJobTreeNode>, profileNamesList: string[]) {
-  // Filter to list of the APIs available for current tree explorer
-  let returnList = profileNamesList.filter((profileName) => {
-      const profile = Profiles.getInstance().loadNamedProfile(profileName);
-      if (zoweNodeProvider instanceof USSTree) {
-          const ussProfileTypes = ZoweExplorerApiRegister.getInstance().registeredUssApiTypes();
-          return ussProfileTypes.includes(profile.type);
-      }
-      if (zoweNodeProvider instanceof DatasetTree) {
-          const mvsProfileTypes = ZoweExplorerApiRegister.getInstance().registeredMvsApiTypes();
-          return mvsProfileTypes.includes(profile.type);
-      }
-      if (zoweNodeProvider instanceof ZosJobsProvider) {
-          const jesProfileTypes = ZoweExplorerApiRegister.getInstance().registeredJesApiTypes();
-          return jesProfileTypes.includes(profile.type);
-      }
-  });
-
-  if (returnList) {
-      returnList = returnList.filter((profileName) =>
-          // Find all cases where a profile is not already displayed
-          !zoweNodeProvider.mSessionNodes.find((sessionNode) => sessionNode.getProfileName() === profileName )
-      );
-  }
-  return returnList;
-}
-
-/**
- * Select a profile regardless of whether we are in Thei or not
- *
- * @param profileNamesList - List of filtered profile names
- */
-async function selectProfile(profileNamesList: string[]) {
-  const createNewProfile = "Create a New Connection to z/OS";
-
-  let chosenProfile = "";
-  const createPick = new FilterDescriptor("\uFF0B " + createNewProfile);
-  const items: vscode.QuickPickItem[] = profileNamesList.map((element) => new FilterItem(element));
-  const quickpick = vscode.window.createQuickPick();
-  const placeholder = localize("addSession.quickPickOption",
-      "Choose \"Create new...\" to define a new profile or select an existing profile to Add to the USS Explorer");
-
-  if (globals.ISTHEIA) {
-      const options: vscode.QuickPickOptions = {
-          placeHolder: placeholder
-      };
-      // get user selection
-      const choice = (await vscode.window.showQuickPick([createPick, ...items], options));
-      if (!choice) {
-          vscode.window.showInformationMessage(localize("enterPattern.pattern", "No selection made."));
-          return;
-      }
-      chosenProfile = choice === createPick ? "" : choice.label;
-  } else {
-      quickpick.items = [createPick, ...items];
-      quickpick.placeholder = placeholder;
-      quickpick.ignoreFocusOut = true;
-      quickpick.show();
-      const choice = await resolveQuickPickHelper(quickpick);
-      quickpick.hide();
-      if (!choice) {
-          vscode.window.showInformationMessage(localize("enterPattern.pattern", "No selection made."));
-          return;
-      }
-      if (choice instanceof FilterDescriptor) {
-          chosenProfile = "";
-      } else {
-          chosenProfile = choice.label;
-      }
-  }
-
-  const quickpickValue = quickpick.value ? quickpick.value : "";
-  return { chosenProfile, quickpickValue };
-}
-
-/**
- * Create a user connection/profile based on the information entered by the user
- *
- * @param zoweNodeProvider - either the USS, MVS, JES tree
- * @param quickpickValue - Name entered by the user
- */
-async function createProfile(zoweNodeProvider: IZoweTree<IZoweDatasetTreeNode|IZoweUSSTreeNode|IZoweJobTreeNode>, quickpickValue: string) {
-  let newprofile: any;
-  let profileName: string;
-  if (quickpickValue) {
-      profileName = quickpickValue;
-  }
-
-  const options = {
-      placeHolder: localize("createNewConnection.option.prompt.profileName.placeholder", "Connection Name"),
-      prompt: localize("createNewConnection.option.prompt.profileName", "Enter a name for the connection"),
-      value: profileName
-  };
-  profileName = await vscode.window.showInputBox(options);
-  if (!profileName) {
-      vscode.window.showInformationMessage(localize("createNewConnection.enterprofileName",
-          "Profile Name was not supplied. Operation Cancelled"));
-      return;
-  }
-  globals.LOG.debug(localize("addSession.log.debug.createNewProfile", "User created a new profile"));
-  try {
-      newprofile = await Profiles.getInstance().createNewConnection(profileName.trim());
-  } catch (error) { errorHandling(error, profileName.trim(), error.message); }
-  if (newprofile) {
-      try {
-          await Profiles.getInstance().refresh();
-      } catch (error) {
-          errorHandling(error, newprofile, error.message);
-      }
-      await zoweNodeProvider.addSession(newprofile);
-      zoweNodeProvider.refresh();
   }
 }
